% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_prep.R
\name{deframe_ranks}
\alias{deframe_ranks}
\title{Extract Named Ranking Vector}
\usage{
deframe_ranks(rank_df, id_type = c("ENTREZID", "SYMBOL"), value = "stat")
}
\arguments{
\item{rank_df}{Data frame from prep_ranks()}

\item{id_type}{Either "ENTREZID" or "SYMBOL"}

\item{value}{Column to use as values (default "stat")}
}
\value{
Named numeric vector
}
\description{
Extract Named Ranking Vector
}
