% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_msigdb.R
\name{run_gsea_msigdb_suite}
\alias{run_gsea_msigdb_suite}
\title{Run Multiple MSigDB Collections}
\usage{
run_gsea_msigdb_suite(
  ranked_symbols,
  collections = c("hallmarks", "reactome"),
  kegg_variant = "KEGG_MEDICUS",
  species = "Homo sapiens",
  parallel = FALSE,
  n_workers = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{ranked_symbols}{Named numeric vector}

\item{collections}{Vector of collection names}

\item{kegg_variant}{Which KEGG variant to use}

\item{species}{Species name}

\item{parallel}{Use parallel processing?}

\item{n_workers}{Number of parallel workers}

\item{...}{Additional parameters}
}
\value{
Named list of gseaResult objects
}
\description{
Run Multiple MSigDB Collections
}
